+++ wp-content/themes/wp-nycc-pb/functions/pb-cycle.php
@@ -36,21 +36,23 @@ function nycc_pbcycle_meta() {
 
 // save the PB Cycle Meta info
 function save_nycc_pbcycle_meta($post_id, $post) {
-  if ( !wp_verify_nonce( $_POST['nycc_pbcycle_meta_noncename'], plugin_basename(__FILE__) )) {
-    return $post->ID;
-  }
-  if ( !current_user_can( 'edit_post', $post->ID ))
-    return $post->ID;
-  $pbcycle_meta['current_pb_cycle'] = $_POST['current_pb_cycle'];
-  foreach ($pbcycle_meta as $key => $value) {
-    if( $post->post_type == 'revision' ) return;
-    $value = implode(',', (array)$value);
-    if(get_post_meta($post->ID, $key, FALSE)) {
-      update_post_meta($post->ID, $key, $value);
-    } else {
-      add_post_meta($post->ID, $key, $value);
+  if ( isset($_POST['nycc_pbcycle_meta_noncename']) ) {
+    if ( !wp_verify_nonce( $_POST['nycc_pbcycle_meta_noncename'], plugin_basename(__FILE__) )) {
+      return $post->ID;
     }
-    if(!$value) delete_post_meta($post->ID, $key);
+    if ( !current_user_can( 'edit_post', $post->ID ))
+      return $post->ID;
+    $pbcycle_meta['current_pb_cycle'] = $_POST['current_pb_cycle'];
+    foreach ($pbcycle_meta as $key => $value) {
+      if( $post->post_type == 'revision' ) return;
+      $value = implode(',', (array)$value);
+      if(get_post_meta($post->ID, $key, FALSE)) {
+        update_post_meta($post->ID, $key, $value);
+      } else {
+        add_post_meta($post->ID, $key, $value);
+      }
+      if(!$value) delete_post_meta($post->ID, $key);
+    }
   }
 }
 add_action('save_post', 'save_nycc_pbcycle_meta', 1, 2);
