+++ wp-content/themes/wp-nycc/assets/functions/districts.php
@@ -40,21 +40,23 @@ function nycc_district_meta() {
 
 // save the District Meta info
 function save_nycc_district_meta($post_id, $post) {
-  if ( !wp_verify_nonce( $_POST['nycc_district_meta_noncename'], plugin_basename(__FILE__) )) {
-    return $post->ID;
-  }
-  if ( !current_user_can( 'edit_post', $post->ID ))
-    return $post->ID;
-  $district_meta['current_member_site'] = $_POST['current_member_site'];
-  foreach ($district_meta as $key => $value) {
-    if( $post->post_type == 'revision' ) return;
-    $value = implode(',', (array)$value);
-    if(get_post_meta($post->ID, $key, FALSE)) {
-      update_post_meta($post->ID, $key, $value);
-    } else {
-      add_post_meta($post->ID, $key, $value);
+  if ( isset($_POST['nycc_district_meta_noncename']) ) {
+    if ( !wp_verify_nonce( $_POST['nycc_district_meta_noncename'], plugin_basename(__FILE__) )) {
+      return $post->ID;
     }
-    if(!$value) delete_post_meta($post->ID, $key);
+    if ( !current_user_can( 'edit_post', $post->ID ))
+      return $post->ID;
+    $district_meta['current_member_site'] = $_POST['current_member_site'];
+    foreach ($district_meta as $key => $value) {
+      if( $post->post_type == 'revision' ) return;
+      $value = implode(',', (array)$value);
+      if(get_post_meta($post->ID, $key, FALSE)) {
+        update_post_meta($post->ID, $key, $value);
+      } else {
+        add_post_meta($post->ID, $key, $value);
+      }
+      if(!$value) delete_post_meta($post->ID, $key);
+    }
   }
 }
 add_action('save_post', 'save_nycc_district_meta', 1, 2);
